requirements.txt
pyTelegramBotAPI==4.4.0
PyGithub==1.55